refactor: move licensing callout to shared folder
