+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/engines/engines_list.test.tsx
@@ -13,9 +13,8 @@ import { mount, shallow } from 'enzyme';
 
 import { Status } from '../../../../../common/types/api';
 
+import { LicensingCallout } from '../../../shared/licensing_callout/licensing_callout';
 import { EnterpriseSearchEnginesPageTemplate } from '../layout/engines_page_template';
-
-import { LicensingCallout } from '../shared/licensing_callout/licensing_callout';
 
 import { EmptyEnginesPrompt } from './components/empty_engines_prompt';
 import { EnginesListTable } from './components/tables/engines_table';
