+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/new_index/method_connector/method_connector.tsx
@@ -17,14 +17,14 @@ import { Status } from '../../../../../../common/types/api';
 import { docLinks } from '../../../../shared/doc_links';
 import { KibanaLogic } from '../../../../shared/kibana';
 import { LicensingLogic } from '../../../../shared/licensing';
+import {
+  LicensingCallout,
+  LICENSING_FEATURE,
+} from '../../../../shared/licensing_callout/licensing_callout';
 import { AddConnectorApiLogic } from '../../../api/connector/add_connector_api_logic';
 
 import { FetchCloudHealthApiLogic } from '../../../api/stats/fetch_cloud_health_api_logic';
 import { NATIVE_CONNECTORS } from '../../search_index/connector/constants';
-import {
-  LicensingCallout,
-  LICENSING_FEATURE,
-} from '../../shared/licensing_callout/licensing_callout';
 import { NewSearchIndexLogic } from '../new_search_index_logic';
 import { NewSearchIndexTemplate } from '../new_search_index_template';
