+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ingest_pipelines/customize_pipeline_item.test.tsx
@@ -14,7 +14,7 @@ import { shallow } from 'enzyme';
 
 import { EuiCallOut, EuiButton } from '@elastic/eui';
 
-import { LicensingCallout } from '../../../shared/licensing_callout/licensing_callout';
+import { LicensingCallout } from '../../../../../shared/licensing_callout/licensing_callout';
 
 import { CopyAndCustomizePipelinePanel } from './customize_pipeline_item';
